Restructured project. Using custom implemented graphs. Added raw files to describe the graph (inside /res/raw). Added camera support. Removed useless POIDetails activity. Managing POIs now happens for every POI picked from a list manually. Results  are not saved yet on file.

+++ android_app/MapManager/app/src/main/java/htw_berlin/de/mapmanager/ManagePOIConnectionsActivity.java
@@ -1,35 +1,54 @@
 package htw_berlin.de.mapmanager;
 
+import android.Manifest;
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.net.Uri;
 import android.os.Environment;
+import android.provider.MediaStore;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.FileProvider;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.SparseArray;
-import android.widget.ExpandableListView;
+import android.view.View;
+import android.widget.Button;
+import android.widget.ListView;
 
+/*
+Graph library
 import org.graphstream.graph.Graph;
 import org.graphstream.graph.Node;
 import org.graphstream.graph.implementations.MultiGraph;
-import org.graphstream.stream.file.FileSink;
 import org.graphstream.stream.file.FileSinkDGS;
 import org.graphstream.stream.file.FileSource;
 import org.graphstream.stream.file.FileSourceDGS;
+*/
+
 
 import java.io.File;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.util.Iterator;
+import java.io.OutputStream;
+import java.util.ArrayList;
 
-public class ManagePOIConnectionsActivity extends AppCompatActivity {
+import htw_berlin.de.mapmanager.graph.Node;
+import htw_berlin.de.mapmanager.graph.TranslatableAdjacencyMatrixGraph;
+
+public class ManagePOIConnectionsActivity extends AppCompatActivity implements ActivityCompat.OnRequestPermissionsResultCallback {
 
     private static final String LOG_TAG = "ManagePOIConnections";
     private static final String GRAPH_NAME = "Alexanderplatz";
-    private static final String NODE_NAME_ATTRIBUTE = "NODE_NAME_ATTRIBUTE";
 
+    private Node parentNode;
+    private static final int TAKE_PHOTO_CODE = 0;
+    private static int picturesCount = 0;
 
-    // more efficient than HashMap for mapping integers to objects, for ui groups of expandablelistview
-    private SparseArray<Group> groups = new SparseArray<Group>();
-    private MultiGraph graph;
+    // user defined, any code that is not used for any other permission request in this activity
+    private static final int CAMERA_PERMISSION_REQUEST_CODE = 1;
+    private ListView listView;
+    private ConnectionListAdapter adapter;
+    private Button capture;
 
 
     @Override
@@ -38,121 +57,160 @@ public class ManagePOIConnectionsActivity extends AppCompatActivity {
         setContentView(R.layout.activity_manage_poiconnections);
 
 
-        File graphFile = getGraphFile();
-        this.graph = readGraphFile(graphFile);
-/*
-        if(this.graph == null){
-            // file does not exist or error occured
-            System.out.println("this.graph is null");
-        }
-        else {*/
-
-            Iterator<Node> nodeIterator = this.graph.getNodeIterator();
-            while(nodeIterator.hasNext()){
-                Node node = nodeIterator.next();
-                String nodeId = node.getId();
-                System.out.printf("Node: id (%s) \n", nodeId);
-            }
-
-        //}
-
-
         /*
          * Get information about the connections
          */
+
         Intent intent = getIntent();
-        String poiId = intent.getStringExtra(MainActivity.EXTRA_MESSAGE_POI_ID);
-        // TODO: using the poiId, get the POI object from the list of nodes in the graph
-        String poiName = "McDonald's";
+        int poiId = intent.getIntExtra(MainActivity.EXTRA_MESSAGE_POI_ID, -1);
+        if(poiId == -1){
+            throw new IllegalArgumentException("The given poiId is invalid: " + (-1));
+        }
+        this.parentNode = MainActivity.graph.getNodeById(poiId);
+
 
         // TODO: get the ArrayList of images of the POI
-        // TODO: get the edges and corresponding nodes (other POIs) that are directly connected to the current POI
 
+        initListView();
 
-        // TODO : get real data
-        createData();
-        ExpandableListView listView = (ExpandableListView) findViewById(R.id.expandableListView);
-        MyExpandableListAdapter adapter = new MyExpandableListAdapter(this, groups);
+        initCamera();
+    }
+
+    @Override
+    protected void onRestart() {
+        super.onRestart();
+
+        // TODO: Save current graph state to file when coming back from the details
+        //saveGraphToFile();
+    }
+
+    private void saveGraphToFile() {
+        throw new UnsupportedOperationException("TODO: implement saveGraphToFile");
+    }
+
+    private void initListView() {
+
+        listView = (ListView) findViewById(R.id.connectionListView);
+        ArrayList<Node> allOtherNodes = new ArrayList<>();
+        allOtherNodes.addAll(MainActivity.graph.getNodes());
+        allOtherNodes.remove(parentNode);
+        adapter = new ConnectionListAdapter(parentNode, allOtherNodes, this);
         listView.setAdapter(adapter);
 
-
-        // TODO remove
-        fakeGraph();
-
     }
 
-    private void fakeGraph() {
+    private void initCamera() {
+        // check and respectively request camera permissions
+        checkCameraPermissions();
 
-        FileSinkDGS output = new FileSinkDGS();
+        // Here, we are making a folder named picFolder to store
+        // pics taken by the camera using this application.
+        final String dir = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES) + "/picFolder/";
+        File newdir = new File(dir);
+        newdir.mkdirs();
 
-        System.setProperty("gs.ui.renderer", "org.graphstream.ui.j2dviewer.J2DGraphRenderer");
-        Graph graph = new MultiGraph(GRAPH_NAME);
-        // create automatically nodes if an     edge containing a new node gets added
-        graph.addNode("A");
-        graph.addNode("B");
-        graph.addNode("C");
+        capture = (Button) findViewById(R.id.btnCapture);
 
-        graph.addEdge("AB", "A", "B");
-        graph.addEdge("BC", "B", "C");
-        graph.addEdge("CA", "C", "A");
+        capture.setOnClickListener(new View.OnClickListener() {
+            public void onClick(View v) {
 
-        writeToFile(graph, output);
+                // Here, the counter will be incremented each time, and the
+                // picture taken by camera will be stored as 1.jpg,2.jpg
+                // and likewise.
+                picturesCount++;
+                String file = dir+ picturesCount +".jpg";
+                File newfile = new File(file);
+                try {
+                    newfile.createNewFile();
+                }
+                catch (IOException e)
+                {
+                    e.printStackTrace();
+                }
+
+                Uri outputFileUri = FileProvider.getUriForFile(ManagePOIConnectionsActivity.this, getApplicationContext().getPackageName() + ".provider", newfile);
+
+                Intent cameraIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
+                cameraIntent.putExtra(MediaStore.EXTRA_OUTPUT, outputFileUri);
+
+                startActivityForResult(cameraIntent, TAKE_PHOTO_CODE);
+            }
+        });
     }
 
-    private void writeToFile(Graph graph, FileSinkDGS output) {
+    private void checkCameraPermissions() {
+        if (checkSelfPermission(Manifest.permission.CAMERA)
+                != PackageManager.PERMISSION_GRANTED) {
+
+            requestPermissions(new String[]{Manifest.permission.CAMERA},
+                    CAMERA_PERMISSION_REQUEST_CODE);
+        }
+    }
 
 
-        if(!isExternalStorageReadable()){
+    public void onRequestPermissionResult(int requestCode, String[] permissions, int[] grantResults) {
+        if (requestCode == CAMERA_PERMISSION_REQUEST_CODE) {
+            if (grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                // Now user should be able to use camera
+            }
+            else {
+                // Your app will not have this permission. Turn off all functions
+                // that require this permission or it will force close
+                capture.setEnabled(false);
+            }
+        }
+    }
+
+    @Override
+    public void onBackPressed() {
+        // TODO: at the moment the changes are saved already when the checkboxes are changed
+        askForSave();
+    }
+
+    /**
+     * Exit the app if user select yes.
+     */
+
+    private void askForSave() {
+
+        AlertDialog.Builder alertDialog = new AlertDialog.Builder(
+                ManagePOIConnectionsActivity.this);
+
+        alertDialog.setPositiveButton("Yes", new DialogInterface.OnClickListener() {
+
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+                finish();
+            }
+        });
+
+        alertDialog.setNegativeButton("No", new DialogInterface.OnClickListener() {
+
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+                finish();
+            }
+        });
+
+        alertDialog.setMessage("Do you want to save the changes before exiting?");
+        alertDialog.setTitle("Save changes?");
+        alertDialog.show();
+    }
+
+
+    private void writeToFile(TranslatableAdjacencyMatrixGraph graph, OutputStream outputStream) {
+
+
+        if (!isExternalStorageReadable()) {
             System.err.println("not readable");
         }
 
-        if(!isExternalStorageWritable()){
+        if (!isExternalStorageWritable()) {
             System.err.println("not writable");
         }
 
-        String filename = "graphData.dgs";
-        //output.begin();?? <-- USe this method?
-        try {
-            File path = getGraphStorageDir();
-            path.mkdirs(); // ensure the directory exists
-
-            // get the file containing the graph
-            File storageFile = getGraphFile();
-
-            System.out.println("file exists: " + storageFile.exists());
-            System.out.println("file name: " + storageFile.getAbsolutePath());
-            System.out.println("file isHidden " + storageFile.isHidden());
-            System.out.println("file is File " + !storageFile.isDirectory());
-            System.out.println("file can write: " + storageFile.canWrite());
-
-            boolean created = storageFile.createNewFile();
-            System.out.println("created? " + created);
-/*
-            FileWriter fw = new FileWriter(storageFile);
-            fw.write("HELLO?");r
-            */
-            FileSinkDGS fileSinkDGS = new FileSinkDGS();
-            fileSinkDGS.writeAll(graph, storageFile.getAbsolutePath());
-
-            //graph.write(output, filename);
-            //output.writeAll(graph, fw);
-            //System.out.println("file exists after writeAll: " + storageFile.exists());
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
     }
-
-    public void createData() {
-        for (int j = 0; j < 5; j++) {
-            Group group = new Group("Test " + j);
-            for (int i = 0; i < 5; i++) {
-                group.children.add("Sub Item" + i);
-            }
-            groups.append(j, group);
-        }
-    }
-
-    /* Checks if external storage is available for read and write */
+    /* Checks if external storage is available for read and write*/
     public boolean isExternalStorageWritable() {
         String state = Environment.getExternalStorageState();
         if (Environment.MEDIA_MOUNTED.equals(state)) {
@@ -161,7 +219,7 @@ public class ManagePOIConnectionsActivity extends AppCompatActivity {
         return false;
     }
 
-    /* Checks if external storage is available to at least read */
+    /* Checks if external storage is available to at least read*/
     public boolean isExternalStorageReadable() {
         String state = Environment.getExternalStorageState();
         if (Environment.MEDIA_MOUNTED.equals(state) ||
@@ -175,50 +233,25 @@ public class ManagePOIConnectionsActivity extends AppCompatActivity {
         // Get the directory for the user's public documents directory.
         // TODO: Warning, files saved in this directory can't be seen from the user itself (from the phone ui)
         return getExternalFilesDir(Environment.DIRECTORY_DOCUMENTS);
-
-        /*if(file.exists()) {
+/*
+        if(file.exists()) {
             file.delete();
         }
-        */
-/*
         if (!file.mkdirs()) {
             Log.e(LOG_TAG, "Directory not created");
         }
-*/
-        /*
+
+
         if(!file.exists()){
             file.createNewFile();
-        }
-*/
+        }*/
+
     }
 
     public File getGraphFile() {
         return new File(this.getGraphStorageDir(), GRAPH_NAME);
     }
 
-    /**
-     * Tutorial:
-     * @link {http://graphstream-project.org/doc/Tutorials/Reading-files-using-FileSource/}
-     * @param graphFile
-     * @return
-     */
-    public MultiGraph readGraphFile(File graphFile){
-        MultiGraph graph = new MultiGraph(GRAPH_NAME) ;
-        FileSource fs = new FileSourceDGS();
-
-        fs.addSink(graph);
-
-        try {
-            fs.readAll(graphFile.getAbsolutePath());
-        } catch( IOException e) {
-            System.err.println("Error reading graph");
-            e.printStackTrace();
-        } finally {
-            fs.removeSink(graph);
-        }
-
-        return graph;
-    }
 
 
 }
